Make the following changes:
,

Find this element:

  <main class="flex-grow pb-20">

with the text content:

    Innovatech SolutionsYour Company Profile SettingsCompany DetailsManage your company's public information.Email: sales@innovatech.comWebsite: www.innovatech.comAddress: 123 Tech Park, Silicon Valley, CALeading provider of innovative tech solutions for businesses. We focus on quality and customer satisfaction, delivering cutting-edge products that drive growth and efficiency. Edit Company DetailsYour Listed Products Manage Orders Add New ProductAdvanced AI Processor UnitState-of-the-art AI processor for demanding applications. Fe...$499.991 unitQuantum Entanglement CommunicatorPrototype device for instantaneous communication across any ...$999999.991 device

Relevant files:

And change it as follows:

noo need of the options edit company details manage orders and add new product her. Just the edit company profile in the setting can be worked out it is fine. And noo need of separate icon in the navbar below with the settings options instead it can be opened with the settings inthe profile page itself