Make the following changes:
,

Find this element:

  <button class="inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 [&amp;_svg]:pointer-events-none [&amp;_svg]:size-4 [&amp;_svg]:shrink-0 bg-destructive text-destructive-foreground hover:bg-destructive/90 h-10 px-4 py-2 w-full sm:w-auto">

with the text content:

    Logout

Relevant files:

And change it as follows:

Now make the logout option when clicke asking a confirmation and then going to the login page if confirmed. The logiin page to be having login with email address and then phone nmber. Then it should redirect to the explore page if the account is alreaddy present inn database. Else there to be a details filling page which ill be having the automatic filling up of emil adress and the contact number, then the filling to be about the company name , gst number.